Add files via upload
Fake news on different platforms is spreading widely and is a matter of serious concern, as it causes social wars and permanent brreakage of the bonds established among people. This project basically consistent of simple Machine Learning Model using Logistic Regression to detect whether a news article is fake or not.